create composable for the CRUD  and start rebuilding the CRUD!
